+++ crates/vm-device/src/device/virtio/virtio_blk.rs
@@ -90,7 +90,7 @@ where
     C: MemoryContainer,
 {
     const NAME: &str = "virtio-blk";
-    const DEVICE_ID: u32 = DeviceId::Blk as u32;
+    const DEVICE_ID: u16 = DeviceId::Blk as u16;
     const DEVICE_FEATURES: u64 = (1 << VIRTIO_F_VERSION_1);
 
     fn virtqueues_size_max(&self) -> Vec<Option<u32>> {
